+++ ximpol/srcmodel/config/sample_single_ptsrc.py
@@ -22,9 +22,9 @@ import numpy
 from ximpol.srcmodel.roi import xPointSource, xROIModel
 from ximpol.srcmodel.spectrum import power_law, constant
 
-ROI_MODEL = xROIModel(90., 0., 1.0)
+ROI_MODEL = xROIModel(90., 0.)
 
-source1 = xPointSource(name='source1', ra=90.0, dec=0.0)
+source1 = xPointSource(name='source1', ra=90., dec=0.)
 source1.spectrum = power_law(10., 2.)
 source1.polarization_degree = constant(0.5)
 source1.polarization_angle = constant(numpy.radians(65.))